requirements.txt
# Core Django
Django>=5.0,<6.0
django-unfold>=0.38.0

# Image processing
Pillow>=10.0.0

# WebSocket support
channels>=4.0.0
channels-redis>=4.1.0

# Redis for caching and WebSocket backend
redis>=5.0.0

# Async support
aioredis>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Type hints support
django-stubs>=4.2.0
types-redis>=4.6.0

# API support
djangorestframework>=3.14.0

# Task queue for async processing
celery>=5.3.0
django-celery-beat>=2.5.0

# Production server
daphne>=4.0.0
